refactor: Move LSP fixture components to LSP fixtures subfolder
